requirements.txt
Flask~=3.1.0
Flask-PyMongo~=2.3.0
pymongo~=4.10.1
Werkzeug~=3.1.3
python-dotenv~=1.0.1
PyJWT~=2.10.1
smtp2go~=2.3.1
python-magic~=0.4.27
pytest~=8.3.4